Tidied up pending client commands
